requirements.txt
streamlit
python-dotenv
langchain_community
langchain_openai
